Cherry-Pick - Move FlattenedAssocCommOp to header for serialization accessibility
